#2 refactoring
