--- app/Http/Controllers/AproposController.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Illuminate\Http\Request;
-
-class AproposController extends Controller
-{
-    public function index()
-    {
-        return view('template/apropos-de-nous');
-    }
-    
-}
