+++ app/Http/Controllers/ong/MaintController.php
@@ -1,0 +1,56 @@
+<?php
+
+namespace App\Http\Controllers\ong;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
+
+/**
+ * in this cntroller we have pilloried the whole application it is  here that we have metered the roads for the reception, who are we, our mission..
+ * 
+ */
+
+class MaintController extends Controller
+{
+    /**
+     * Display home page.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        //
+    }
+
+    /**
+     * Show all missions link.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function orMission()
+    {
+        //
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function aboutUs()
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function howToHelp($id)
+    {
+        //
+    }
+
+}
