+++ app/Models/Don.php
@@ -5,7 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Maladie extends Model
+class Don extends Model
 {
     use HasFactory;
 }
