+++ app/Models/Member.php
@@ -5,7 +5,7 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class FaireUnDon extends Model
+class Member extends Model
 {
     use HasFactory;
 }
